day 07: able to show products, add to cart, & checkout order

+++ Task/MiniStore/src/com/dansmultipro/ministore/model/Product.java
@@ -28,7 +28,7 @@ public class Product {
         return price;
     }
 
-    public double getStock() {
+    public int getStock() {
         return stock;
     }
